Initial Commit of KPIC-VFN Simulator
This is the first functional version of the KPIC-VFN on-sky simulator. It is heavily based on the PolychromaticSimTests script but enables the use of real WF data. It also implements several performance improvements. 
This script can still produce synthetic wavefronts so it can do everything that the old (Poly...SimTests) script could do.