Commit various sim scripts for testing VFN performance
The new MaskManufacturing_Simulations directory has scripts for simulating the performance of VFN given different mask manufacturing parameters. There are two scripts for simulating amplitude+phase defects in the pupil as would be introduced by a BeamCo mask if we reduce the manufacturing specifications. There is also a script for testing if a Zeiss-like step vortex would work for VFN.

The KPICVFN_Simulator directory has 3 new scripts:
- pupilAreaComparator is used for computing losses due to different pupil areas in KPIC VFN vs DS mode.
- CouplingWithWFE simulates coupling efficiencies in the presence of a "semi-realistic" PSD of WFR from an AO system. 
- IntegrationTimeComparator uses eta_s/eta_p^2 as derived by G.Ruane 2018 to make arguments about the separations at which VFN can be expected to provide integration time reductions over the DS mode.